requirements.txt
moviepy>=1.0.3
openai-whisper>=20231117